Cleanup templates p1
